feat(dependencies): add getDependencies function for synthetic dependency graph generation refactor(types): update type checking for check:types script with increased memory limit chore(tests): add storybook example for Table component using getDependencies
