+++ ui/package.json
@@ -13,7 +13,7 @@
         "build:design": "npm run build --workspace=@kestra-io/design-system",
         "build:topology": "npm run build --workspace=@kestra-io/topology",
         "preview": "vite preview",
-        "check:types": "vue-tsc --build tsconfig.app.json && echo \"app checked\" && vue-tsc --project tsconfig.test.json && echo \"test checked\"",
+        "check:types": "NODE_OPTIONS=--max-old-space-size=4096 vue-tsc --build tsconfig.app.json && echo \"app checked\" && NODE_OPTIONS=--max-old-space-size=4096 vue-tsc --project tsconfig.test.json && echo \"test checked\"",
         "test:lint": "oxlint src packages && eslint",
         "test:all": "vitest run --coverage",
         "test:unit": "vitest run --project=unit",
